split cloud/platform && license module
**Phenomenon and reproduction steps**

**Root cause and solution**

- easy to module replace

**Impactions**

**Test method**

- make success

**Affected branch(es)**

* main

**Checklist**

- [ ] Dependencies update required
- [ ] Common bug (similar problem in other repo)
